Added router and first site structure
