+++ frontend/src/App.css
@@ -32,8 +32,11 @@ html {
   margin: 1rem;
 }
 
-.btn {
+.frontPageBtn {
   margin-top: 10rem;
+}
+
+.btn {
   padding: 1rem;
   min-width: 6rem; 
   background-color: lightgreen;
