+++ frontend/src/App.js
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
 import { BrowserRouter, Route, Link } from 'react-router-dom';
 import queryString from 'query-string';
+import MainPage from './components/MainPage';
 import Login from './components/Login';
 import SpotifyLogin from './components/SpotifyLogin';
 import "./App.css";
+import seleqt from './images/seleqt.png';
 
 
 class App extends Component {
@@ -46,11 +48,12 @@ class App extends Component {
     return (
       <BrowserRouter>
         <div className="App">
-          <Login />
-          <SpotifyLogin />
-           <div>
+           {/* <div>
              <Link  to="/firstTimeUser"><button> Test click </button> </ Link>
-           </div>
+           </div> */}
+           <Route exact path="/" component={Login} />
+           <Route path="/firstTimeUser" component={SpotifyLogin} />
+           <Route path="/mainPage" component={MainPage} />
         </div>
       </BrowserRouter>
     );
